🔨 Refactor: QueryService와 CommandService 분리
